Remove or hardcode CVS $Date$ and $Id$ tags
When these files were kept in W3C's CVS repository CVS would update
these tags upon each commit, but now that they are managed on GitHub
that expansion will no longer take place. This commit removes or
hardcodes the $Date$ and $Id$ tags to the date/time of the last commit
in the CVS repo. They can be updated manually as needed in the future.